backend_upload_video_code/requirements.txt
asgiref==3.8.1
boto3==1.35.74
botocore==1.35.74
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jmespath==1.0.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
urllib3==2.2.3
